% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fonction1}
\alias{fonction1}
\title{Nom de la fonction}
\usage{
fonction1(x)
}
\arguments{
\item{x}{definition param x}

\item{z}{definition param z}
}
\description{
Description de la fonction
}
\examples{
\dontrun{

un exemple de fonction


}
}
