% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fonctionx}
\alias{fonctionx}
\title{Nom de la fonction 2}
\usage{
fonctionx(x)
}
\arguments{
\item{x}{definition param x}

\item{z}{definition param x}
}
\description{
Description de la fonction une autre donction
}
\details{
Additional details... ce que tu veux
}
\note{
dmjzdfjsgjsg

sdhshqd
}
\seealso{
titit totot
}
